retain backward compatibility with Go 1.0.x
Go 1.0.x doesn't have json.Number
