+++ simplejson.go
@@ -1,7 +1,6 @@
 package simplejson
 
 import (
-	"bytes"
 	"encoding/json"
 	"errors"
 	"log"
@@ -30,14 +29,6 @@ func NewJson(body []byte) (*Json, error) {
 // Encode returns its marshaled data as `[]byte`
 func (j *Json) Encode() ([]byte, error) {
 	return j.MarshalJSON()
-}
-
-// Implements the json.Unmarshaler interface.
-func (j *Json) UnmarshalJSON(p []byte) error {
-	dec := json.NewDecoder(bytes.NewReader(p))
-	dec.UseNumber()
-
-	return dec.Decode(&j.data)
 }
 
 // Implements the json.Marshaler interface.
@@ -153,33 +144,6 @@ func (j *Json) String() (string, error) {
 		return s, nil
 	}
 	return "", errors.New("type assertion to string failed")
-}
-
-// Float64 type asserts to `json.Number` then converts to `float64`
-func (j *Json) Float64() (float64, error) {
-	if n, ok := (j.data).(json.Number); ok {
-		return n.Float64()
-	}
-	return -1, errors.New("type assertion to float64 failed")
-}
-
-// Int type asserts to `json.Number` then converts to `int`
-func (j *Json) Int() (int, error) {
-	if n, ok := (j.data).(json.Number); ok {
-		i, ok := n.Int64()
-		return int(i), ok
-	}
-
-	return -1, errors.New("type assertion to float64 failed")
-}
-
-// Int type asserts to `json.Number` then converts to `int64`
-func (j *Json) Int64() (int64, error) {
-	if n, ok := (j.data).(json.Number); ok {
-		return n.Int64()
-	}
-
-	return -1, errors.New("type assertion to float64 failed")
 }
 
 // Bytes type asserts to `[]byte`
